requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# PDF Processing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# Word Processing
python-docx>=1.1.0

# LLM
litellm>=1.30.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
